refactor(split-meow): HistoryTab 抽出 EditExpenseSheet 與 SettlementSection
- HistoryTab 1197 行降至 701 行：編輯 sheet 與結算域各自成檔，列表職責單一化
- EditExpenseSheet.tsx：編輯消費 bottom sheet 純搬移（含 z-[60] 層級與 store selector 註解）
- SettlementSection.tsx：結清方式與各人結算兩區塊純搬移，expandedMemberId 狀態隨遷
- calculateSettlements 落 lib/settlements.ts，避免元件檔混出工具函式（react-refresh 規範）
- 行為零變更：JSX、狀態語意與 store 讀取完全一致，既有測試無需修改

測試：pnpm --filter @app/split-meow test:run 257 綠、typecheck、lint 0 warning 通過
