+++ apps/split-meow/src/components/HistoryTab.tsx
@@ -4,7 +4,6 @@ import { useStore, type Member, type ExpenseRecord } from '../store/useStore';
 import {
   convertAmount,
   formatAmount,
-  getCurrencySymbol,
   resolveExpenseCurrency,
   resolveTripCurrency,
   isMixedCurrencyTrip,
@@ -14,47 +13,9 @@ import { format } from 'date-fns';
 import { useEffect, useRef, useState } from 'react';
 import { cn } from '../lib/utils';
 import { MemberAvatar } from './MemberAvatar';
-
-interface Settlement {
-  from: string;
-  to: string;
-  amount: number;
-}
-
-function calculateSettlements(balances: Record<string, number>): Settlement[] {
-  const creditors = Object.entries(balances)
-    .filter(([, b]) => b > 0.01)
-    .map(([id, amount]) => ({ id, amount }))
-    .sort((a, b) => b.amount - a.amount);
-
-  const debtors = Object.entries(balances)
-    .filter(([, b]) => b < -0.01)
-    .map(([id, amount]) => ({ id, amount: -amount }))
-    .sort((a, b) => b.amount - a.amount);
-
-  const settlements: Settlement[] = [];
-  let ci = 0;
-  let di = 0;
-
-  while (ci < creditors.length && di < debtors.length) {
-    const credit = creditors[ci];
-    const debt = debtors[di];
-    if (!credit || !debt) break;
-    const amount = Math.min(credit.amount, debt.amount);
-
-    if (amount > 0.01) {
-      settlements.push({ from: debt.id, to: credit.id, amount });
-    }
-
-    credit.amount -= amount;
-    debt.amount -= amount;
-
-    if (credit.amount < 0.01) ci++;
-    if (debt.amount < 0.01) di++;
-  }
-
-  return settlements;
-}
+import { EditExpenseSheet } from './EditExpenseSheet';
+import { SettlementSection, BalancesSection } from './SettlementSection';
+import { calculateSettlements } from '../lib/settlements';
 
 /** 最多顯示 MAX_AVATAR 個頭像，其餘以 +N 呈現 */
 const MAX_AVATAR = 3;
@@ -138,7 +99,6 @@ export function HistoryTab() {
   const updateExpense = useStore((s) => s.updateExpense);
   const toggleSettlement = useStore((s) => s.toggleSettlement);
   const [expandedId, setExpandedId] = useState<string | null>(null);
-  const [expandedMemberId, setExpandedMemberId] = useState<string | null>(null);
   const [editingNoteId, setEditingNoteId] = useState<string | null>(null);
   const [editingNoteValue, setEditingNoteValue] = useState('');
   const noteInputRef = useRef<HTMLInputElement>(null);
@@ -362,296 +322,24 @@ export function HistoryTab() {
 
       {/* 結清方式 */}
       {settlements.length > 0 && !isMixedCurrency && (
-        <section className="mb-10">
-          <h3 className="text-xs font-medium uppercase tracking-widest text-outline px-2 mb-4">
-            {t('history.settlements')}
-          </h3>
-          <div className="space-y-3">
-            {settlements.map((s, i) => {
-              const fromMember = members.find((m) => m.id === s.from);
-              const toMember = members.find((m) => m.id === s.to);
-              if (!fromMember || !toMember) return null;
-              const settlementKey = `${currentTripId ?? ''}:${s.from}:${s.to}`;
-              const isSettled = settledPayments.includes(settlementKey);
-              return (
-                <div
-                  key={i}
-                  data-testid="settlement-row"
-                  role="button"
-                  tabIndex={0}
-                  aria-pressed={isSettled}
-                  onClick={() => toggleSettlement(settlementKey)}
-                  onKeyDown={(e) => {
-                    if (e.key === 'Enter' || e.key === ' ') {
-                      e.preventDefault();
-                      toggleSettlement(settlementKey);
-                    }
-                  }}
-                  className={cn(
-                    'bg-surface-container-lowest p-4 rounded-[1.5rem] flex items-center gap-3 shadow-ambient cursor-pointer transition-opacity active:scale-[0.98] transition-transform',
-                    isSettled && 'opacity-50',
-                  )}
-                >
-                  {/* 誰付誰方向視覺化：付款人 → 收款人 */}
-                  <div className="flex items-center gap-1 shrink-0">
-                    <MemberAvatar seed={fromMember.avatarUrl} alt={fromMember.name} size={36} />
-                    <span
-                      className="material-symbols-outlined text-[18px] text-primary"
-                      aria-hidden="true"
-                    >
-                      arrow_forward
-                    </span>
-                    <MemberAvatar seed={toMember.avatarUrl} alt={toMember.name} size={36} />
-                  </div>
-                  <div className="flex-1 min-w-0">
-                    <p
-                      className={cn(
-                        'text-sm text-on-surface-variant leading-snug',
-                        isSettled && 'line-through',
-                      )}
-                    >
-                      {t('history.pay_to', {
-                        from: fromMember.name,
-                        to: toMember.name,
-                      })}
-                    </p>
-                    {isSettled && (
-                      <p className="text-xs text-tertiary font-medium mt-0.5">
-                        {t('history.balance_settled')}
-                      </p>
-                    )}
-                  </div>
-                  <div
-                    className={cn(
-                      'rounded-full px-3 py-1 shrink-0 transition-colors',
-                      isSettled
-                        ? 'bg-tertiary-container text-on-tertiary-container'
-                        : 'bg-secondary-container text-on-secondary-container',
-                    )}
-                  >
-                    <span className="text-sm font-bold">
-                      {isSettled ? (
-                        <span
-                          className="material-symbols-outlined text-[16px] leading-none"
-                          style={{ fontVariationSettings: "'FILL' 1" }}
-                        >
-                          check_circle
-                        </span>
-                      ) : (
-                        formatAmount(s.amount, tripCurrency)
-                      )}
-                    </span>
-                  </div>
-                </div>
-              );
-            })}
-          </div>
-        </section>
+        <SettlementSection
+          settlements={settlements}
+          members={members}
+          currentTripId={currentTripId}
+          tripCurrency={tripCurrency}
+          settledPayments={settledPayments}
+          onToggle={toggleSettlement}
+        />
       )}
 
       {/* 各人結算 */}
       {!isMixedCurrency && Object.keys(balances).length > 0 && (
-        <section className="mb-10">
-          <h3 className="text-xs font-medium uppercase tracking-widest text-outline px-2 mb-4">
-            {t('history.balances')}
-          </h3>
-          <div className="space-y-3">
-            {Object.entries(balances).map(([memberId, amount]) => {
-              const member = members.find((m) => m.id === memberId);
-              if (!member) return null;
-              const isOwed = amount > 0;
-              const isSettled = Math.abs(amount) < 0.01;
-              const isOpen = expandedMemberId === memberId;
-
-              // 與此成員相關的消費（有參與的）
-              const memberExpenses = tripExpenses.filter(
-                (e) => e.paidBy === memberId || memberId in e.perPersonAmounts,
-              );
-
-              return (
-                <div
-                  key={memberId}
-                  role={isSettled ? undefined : 'button'}
-                  tabIndex={isSettled ? undefined : 0}
-                  aria-expanded={isSettled ? undefined : isOpen}
-                  className={cn(
-                    'bg-surface-container-lowest rounded-[1.5rem] shadow-ambient overflow-hidden transition-all duration-300',
-                    !isSettled && 'cursor-pointer',
-                  )}
-                  onClick={() => {
-                    if (isSettled) return;
-                    setExpandedMemberId(isOpen ? null : memberId);
-                  }}
-                  onKeyDown={(e) => {
-                    if (isSettled || e.target !== e.currentTarget) return;
-                    if (e.key === 'Enter' || e.key === ' ') {
-                      e.preventDefault();
-                      setExpandedMemberId(isOpen ? null : memberId);
-                    }
-                  }}
-                >
-                  {/* 摘要列 */}
-                  <div className="p-4 flex items-center gap-3">
-                    <MemberAvatar seed={member.avatarUrl} alt={member.name} size={40} />
-                    <div className="flex-1 min-w-0">
-                      <p className="font-medium text-sm text-on-surface truncate">{member.name}</p>
-                      {isSettled ? (
-                        <p className="text-xs text-on-surface-variant/60">
-                          {t('history.balance_settled')}
-                        </p>
-                      ) : (
-                        <p
-                          className={cn(
-                            'text-xs font-medium',
-                            isOwed ? 'text-secondary' : 'text-error',
-                          )}
-                        >
-                          {isOwed ? t('history.owed') : t('history.owing')}
-                          {' · '}
-                          {t('history.tap_to_detail')}
-                        </p>
-                      )}
-                    </div>
-                    {!isSettled && (
-                      <div className="flex items-center gap-2 shrink-0">
-                        <span
-                          className={cn(
-                            'text-sm font-bold px-3 py-1 rounded-full',
-                            isOwed
-                              ? 'bg-secondary-container text-on-secondary-container'
-                              : 'bg-error-container text-on-error-container',
-                          )}
-                        >
-                          {isOwed ? '+' : '-'}
-                          {formatAmount(Math.abs(amount), tripCurrency)}
-                        </span>
-                        <span
-                          className="material-symbols-outlined text-on-surface-variant text-lg transition-transform duration-300"
-                          style={{ transform: isOpen ? 'rotate(180deg)' : 'rotate(0deg)' }}
-                        >
-                          expand_more
-                        </span>
-                      </div>
-                    )}
-                  </div>
-
-                  {/* 展開明細；收合時 inert 消除幻影可聚焦項 */}
-                  <div
-                    inert={!isOpen || undefined}
-                    className={cn(
-                      'grid transition-all duration-200 ease-in-out',
-                      isOpen ? 'grid-rows-[1fr] opacity-100' : 'grid-rows-[0fr] opacity-0',
-                    )}
-                  >
-                    <div className="overflow-hidden">
-                      <div
-                        className="px-4 pb-4 space-y-2 border-t border-outline-variant/15 pt-3"
-                        onClick={(e) => e.stopPropagation()}
-                      >
-                        {memberExpenses.map((exp, idx) => {
-                          const isPayer = exp.paidBy === memberId;
-                          const share = exp.perPersonAmounts[memberId] ?? 0;
-                          const net = (isPayer ? exp.totalAmount : 0) - share;
-                          const expLabel =
-                            exp.note ||
-                            (exp.type === 'split_evenly'
-                              ? t('history.split_evenly')
-                              : t('history.itemized'));
-
-                          return (
-                            <div
-                              key={exp.id}
-                              className={cn(
-                                'rounded-2xl bg-surface-container-low p-3 space-y-2',
-                                idx > 0 && '',
-                              )}
-                            >
-                              {/* 消費標題 */}
-                              <div className="flex items-center justify-between">
-                                <span className="text-xs font-medium text-on-surface truncate max-w-[140px]">
-                                  {expLabel}
-                                </span>
-                                <span className="text-[10px] text-on-surface-variant/60 shrink-0">
-                                  {format(exp.createdAt, 'M/d HH:mm')}
-                                </span>
-                              </div>
-
-                              {/* 墊付行（只有付款人才有） */}
-                              {isPayer && (
-                                <div className="flex items-center justify-between text-xs">
-                                  <span className="flex items-center gap-1.5 text-secondary">
-                                    <span className="material-symbols-outlined text-[13px]">
-                                      payments
-                                    </span>
-                                    {t('history.balance_paid')}
-                                  </span>
-                                  <span className="font-semibold text-secondary">
-                                    +{formatAmount(exp.totalAmount, expenseCurrency(exp))}
-                                  </span>
-                                </div>
-                              )}
-
-                              {/* 分攤行 */}
-                              {share > 0.01 && (
-                                <div className="flex items-center justify-between text-xs">
-                                  <span className="flex items-center gap-1.5 text-on-surface-variant">
-                                    <span className="material-symbols-outlined text-[13px]">
-                                      group
-                                    </span>
-                                    {t('history.balance_share')}
-                                  </span>
-                                  <span className="font-semibold text-error">
-                                    -{formatAmount(share, expenseCurrency(exp))}
-                                  </span>
-                                </div>
-                              )}
-
-                              {/* 小計 */}
-                              <div className="flex items-center justify-between text-xs border-t border-outline-variant/20 pt-1.5">
-                                <span className="text-on-surface-variant/70">
-                                  {t('history.balance_net')}
-                                </span>
-                                <span
-                                  className={cn(
-                                    'font-bold',
-                                    net > 0.01
-                                      ? 'text-secondary'
-                                      : net < -0.01
-                                        ? 'text-error'
-                                        : 'text-on-surface-variant',
-                                  )}
-                                >
-                                  {net > 0.01 ? '+' : net < -0.01 ? '-' : ''}
-                                  {formatAmount(Math.abs(net), expenseCurrency(exp))}
-                                </span>
-                              </div>
-                            </div>
-                          );
-                        })}
-
-                        {/* 合計 */}
-                        <div className="flex items-center justify-between pt-1 px-1">
-                          <span className="text-xs font-medium text-on-surface-variant uppercase tracking-wider">
-                            {t('history.balance_total')}
-                          </span>
-                          <span
-                            className={cn(
-                              'text-base font-bold',
-                              isOwed ? 'text-secondary' : 'text-error',
-                            )}
-                          >
-                            {isOwed ? '+' : '-'}
-                            {formatAmount(Math.abs(amount), tripCurrency)}
-                          </span>
-                        </div>
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              );
-            })}
-          </div>
-        </section>
+        <BalancesSection
+          balances={balances}
+          members={members}
+          tripExpenses={tripExpenses}
+          tripCurrency={tripCurrency}
+        />
       )}
 
       {/* 近期活動 */}
@@ -1009,188 +697,6 @@ export function HistoryTab() {
           </button>
         </div>
       )}
-    </div>
-  );
-}
-
-// ── EditExpenseSheet ─────────────────────────────────────────────────────────
-
-interface EditExpenseSheetProps {
-  expense: ExpenseRecord;
-  members: Member[];
-  onSave: (
-    patch: Partial<
-      Pick<ExpenseRecord, 'totalAmount' | 'perPersonAmounts' | 'participantIds' | 'paidBy'>
-    >,
-  ) => void;
-  onClose: () => void;
-}
-
-function EditExpenseSheet({ expense, members, onSave, onClose }: EditExpenseSheetProps) {
-  const { t } = useTranslation();
-  const globalCurrency = useStore((s) => s.currency);
-  // selector 不可回傳新陣列（getSnapshot 不穩定會觸發無限迴圈）；filter 於 render 內派生。
-  const allExpenses = useStore((s) => s.expenses);
-  const currentTripId = useStore((s) => s.currentTripId);
-  const tripExpenses = allExpenses.filter((e) => e.tripId === currentTripId);
-  const tripCurrency = resolveTripCurrency(tripExpenses, globalCurrency);
-  const expenseCurrency = resolveExpenseCurrency(expense, tripCurrency);
-  const isEvenly = expense.type === 'split_evenly';
-
-  const [totalInput, setTotalInput] = useState(
-    isEvenly ? String(Math.round(expense.totalAmount)) : '',
-  );
-
-  const [perPersonInputs, setPerPersonInputs] = useState<Record<string, string>>(() => {
-    const init: Record<string, string> = {};
-    Object.entries(expense.perPersonAmounts).forEach(([id, amt]) => {
-      init[id] = String(Math.round(amt));
-    });
-    return init;
-  });
-
-  const [editPayerId, setEditPayerId] = useState(expense.paidBy);
-
-  const handleSave = () => {
-    if (isEvenly) {
-      const total = parseFloat(totalInput);
-      if (!Number.isFinite(total) || total <= 0) return;
-      const participantIds = expense.participantIds;
-      const split = total / participantIds.length;
-      const perPersonAmounts: Record<string, number> = {};
-      participantIds.forEach((id) => {
-        perPersonAmounts[id] = split;
-      });
-      onSave({ totalAmount: total, perPersonAmounts, paidBy: editPayerId });
-    } else {
-      const perPersonAmounts: Record<string, number> = {};
-      let total = 0;
-      Object.entries(perPersonInputs).forEach(([id, val]) => {
-        const n = parseFloat(val);
-        if (Number.isFinite(n) && n > 0) {
-          perPersonAmounts[id] = n;
-          total += n;
-        }
-      });
-      if (total <= 0) return;
-      const participantIds = Object.keys(perPersonAmounts);
-      onSave({ totalAmount: total, perPersonAmounts, participantIds, paidBy: editPayerId });
-    }
-  };
-
-  return (
-    // modal 層級必須高於 BottomNav（z-50），否則同 z 下 DOM 較晚的 nav 會攔截 CTA 點擊（G3）。
-    <div
-      className="fixed inset-0 z-[60] flex items-end"
-      onClick={(e) => {
-        if (e.target === e.currentTarget) onClose();
-      }}
-    >
-      <div className="absolute inset-0 bg-black/40" onClick={onClose} />
-      <div
-        className="relative w-full bg-surface-container-low rounded-t-[2rem] p-6 animate-in slide-in-from-bottom-4 duration-300 space-y-4"
-        style={{ paddingBottom: 'calc(var(--chrome-bottom) + 1rem)' }}
-      >
-        <div className="flex items-center justify-between">
-          <h2 className="text-base font-semibold text-on-surface">{t('history.edit_title')}</h2>
-          <button
-            onClick={onClose}
-            aria-label={t('common.close')}
-            className="w-11 h-11 -mr-2 flex items-center justify-center text-on-surface-variant hover:text-on-surface transition-colors cursor-pointer"
-          >
-            <span className="material-symbols-outlined text-[20px]" aria-hidden="true">
-              close
-            </span>
-          </button>
-        </div>
-
-        {isEvenly ? (
-          <div className="flex items-center gap-3 bg-surface-container rounded-2xl px-4 py-3">
-            <span className="text-on-surface-variant text-sm font-medium shrink-0">
-              {getCurrencySymbol(expenseCurrency)}
-            </span>
-            <input
-              type="number"
-              inputMode="decimal"
-              value={totalInput}
-              onChange={(e) => setTotalInput(e.target.value)}
-              className="flex-1 bg-transparent text-lg font-semibold text-on-surface outline-none"
-              autoFocus
-            />
-          </div>
-        ) : (
-          <div className="space-y-2">
-            {expense.participantIds.map((id) => {
-              const m = members.find((x) => x.id === id);
-              if (!m) return null;
-              return (
-                <div
-                  key={id}
-                  className="flex items-center gap-3 bg-surface-container rounded-2xl px-4 py-2.5"
-                >
-                  <MemberAvatar seed={m.avatarUrl} alt={m.name} size={28} />
-                  <span className="flex-1 text-sm font-medium text-on-surface truncate">
-                    {m.name}
-                  </span>
-                  <span className="text-on-surface-variant text-sm shrink-0">
-                    {getCurrencySymbol(expenseCurrency)}
-                  </span>
-                  <input
-                    type="number"
-                    inputMode="decimal"
-                    value={perPersonInputs[id] ?? ''}
-                    onChange={(e) =>
-                      setPerPersonInputs((prev) => ({ ...prev, [id]: e.target.value }))
-                    }
-                    className="w-20 bg-transparent text-sm font-semibold text-on-surface outline-none text-right"
-                  />
-                </div>
-              );
-            })}
-          </div>
-        )}
-
-        <div>
-          <p className="text-xs font-medium text-on-surface-variant mb-2">
-            {t('history.edit_payer')}
-          </p>
-          <div className="flex flex-wrap gap-2">
-            {members
-              .filter((m) => expense.participantIds.includes(m.id))
-              .map((m) => (
-                <button
-                  key={m.id}
-                  onClick={() => setEditPayerId(m.id)}
-                  aria-pressed={editPayerId === m.id}
-                  className={cn(
-                    'flex items-center gap-1.5 min-h-11 px-3 py-1.5 rounded-full text-sm transition-all cursor-pointer',
-                    editPayerId === m.id
-                      ? 'bg-primary text-on-primary shadow-sm'
-                      : 'bg-surface-container text-on-surface-variant hover:bg-surface-container-high',
-                  )}
-                >
-                  <MemberAvatar seed={m.avatarUrl} alt={m.name} size={20} />
-                  {m.name}
-                </button>
-              ))}
-          </div>
-        </div>
-
-        <div className="flex gap-3 pt-1">
-          <button
-            onClick={onClose}
-            className="flex-1 py-3 rounded-2xl bg-surface-container text-on-surface-variant text-sm font-medium transition-colors hover:bg-surface-container-high cursor-pointer"
-          >
-            {t('history.cancel')}
-          </button>
-          <button
-            onClick={handleSave}
-            className="flex-1 py-3 rounded-2xl bg-primary text-on-primary text-sm font-semibold transition-colors active:scale-95 cursor-pointer"
-          >
-            {t('history.save')}
-          </button>
-        </div>
-      </div>
     </div>
   );
 }
